requirements.txt
python-telegram-bot==22.1
python-dotenv==1.1.0
firebase-admin==6.9.0
